move database to shared
