--- beacon-chain/database/BUILD.bazel
@@ -1,20 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["database.go"],
-    importpath = "github.com/prysmaticlabs/prysm/beacon-chain/database",
-    visibility = ["//beacon-chain:__subpackages__"],
-    deps = [
-        "//shared/database:go_default_library",
-        "@com_github_ethereum_go_ethereum//ethdb:go_default_library",
-        "@com_github_sirupsen_logrus//:go_default_library",
-    ],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["database_test.go"],
-    embed = [":go_default_library"],
-    deps = ["@com_github_syndtr_goleveldb//leveldb/errors:go_default_library"],
-)
