+++ beacon-chain/node/BUILD.bazel
@@ -7,12 +7,12 @@ go_library(
     visibility = ["//beacon-chain:__subpackages__"],
     deps = [
         "//beacon-chain/blockchain:go_default_library",
-        "//beacon-chain/database:go_default_library",
         "//beacon-chain/powchain:go_default_library",
         "//beacon-chain/sync:go_default_library",
         "//beacon-chain/utils:go_default_library",
         "//shared:go_default_library",
         "//shared/cmd:go_default_library",
+        "//shared/database:go_default_library",
         "//shared/debug:go_default_library",
         "//shared/p2p:go_default_library",
         "@com_github_ethereum_go_ethereum//common:go_default_library",
