+++ shared/database/inmemory.go
@@ -6,10 +6,7 @@ import (
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/ethdb"
-	"github.com/sirupsen/logrus"
 )
-
-var log = logrus.WithField("prefix", "db")
 
 // KVStore is an in-memory mapping of hashes to RLP encoded values.
 type KVStore struct {
